Create TempPWDGen.sql
This script creates a stored procedure for generating temporary passwords when users initiate a password reset. I no longer implement this method of password generation, as it's obsolete, insecure, and does not comply with the NIST 800-63 Digital Identity Guidelines standards. As such this is posted strictly for proof of concept purposes.